+++ VM/Lexer/byte_token.hpp
@@ -32,7 +32,7 @@ class ByteToken {
         bool isMemory() const;
 
         ByteToken::Type getType() const;
-        double getValueIfExisting() const;
+        int8_t getValueIfExisting() const;
         std::string getStringValueIfExisting() const;
         int64_t getIntegerValueIfExisting() const;
         double getDoubleValueIfExisting() const;
@@ -43,7 +43,7 @@ class ByteToken {
     protected:
         ByteToken::Type m_type;
 
-        double m_value;
+        int8_t m_value;
         std::string m_string_value;
         int64_t m_integer_value;
         double m_double_value;
